// js_modules/dagit/packages/core/src/runs/types/ComputeLogContentFileFragment.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: ComputeLogContentFileFragment
// ====================================================

export interface ComputeLogContentFileFragment {
  __typename: "ComputeLogFile";
  path: string;
  cursor: number;
  data: string | null;
  downloadUrl: string | null;
}